pyproject.toml
[project]
name = "growzucchini"
version = "0.1.0"
description = "Automated plant growth controller"
authors = [
    { name = "Alexander Suvorov", email = "hi.alexsuv@gmail.com" }
]
dependencies = [
    "pyserial-asyncio>=0.6"
]

[tool.poetry]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"